+++ src/queries.rs
@@ -16,6 +16,8 @@ pub const SWIFT_TAGS_QUERY: &str = include_str!("../queries/swift.scm");
 
 pub const DART_TAGS_QUERY: &str = include_str!("../queries/dart.scm");
 
+pub const KOTLIN_TAGS_QUERY: &str = include_str!("../queries/kotlin.scm");
+
 pub const TERRAFORM_TAGS_QUERY: &str = include_str!("../queries/terraform.scm");
 
 pub const OBJECTIVE_C_TAGS_QUERY: &str = include_str!("../queries/objective_c.scm");