+++ tests/test_cases/kotlin/basic_query/input/script.kts
@@ -1,0 +1,2 @@
+val scriptValue = 1
+fun scriptFunction() = scriptValue