+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Allow users to manage the downloading of WASM grammars not bundled with treetags
 - Add basic tags support for Swift
 - Add basic tags support for Objective C
 - Add basic tags support for Dart
+- Add basic tags support for Kotlin
 
 ### Changed
 - Tag Address is truncated to 96 chars for native parsers